\name{comprdata}
\alias{comprdata}
\docType{data}
\title{
An example dataset with competing risks outcomes.
}
\description{
An example dataset with competing risks outcomes that can be used for \code{comprSensitivity}.
}
\usage{data("comprdata")}
\format{
  The format is a list of 5, corresponding to \code{t}, \code{d}, \code{Z}, \code{X}, \code{U}, respectively.
}
\references{
Huang, R., Xu, R., & Dulai, P. S. (2019). Sensitivity Analysis of Treatment Effect to Unmeasured Confounding in Observational Studies with Survival and Competing Risks Outcomes. arXiv preprint arXiv:1908.01444.
}
\examples{
data(comprdata)
}
\keyword{datasets}
